+++ tests/Transformers/MultibyteTextNormalizerTest.php
@@ -31,9 +31,9 @@ class MultibyteTextNormalizerTest extends TestCase
         $this->dataset = Unlabeled::quick([
             ['The quick brown fox jumped over the lazy man sitting at a bus'
                 . ' stop drinking a can of Coke'],
-            ['with a Dandy   umbrella '],
-            ['Depuis qu’il   avait emménagé à côté de chez elle, il y a de ça cinq ans. '],
-            ['   Working   with emoji 🤓 '],
+            ['with a Dandy   umbrella'],
+            ['Depuis qu’il avait emménagé à côté de chez elle, il y a de ça cinq ans.'],
+            ['Working with emoji 🤓'],
         ]);
 
         $this->transformer = new MultibyteTextNormalizer();
@@ -58,7 +58,7 @@ class MultibyteTextNormalizerTest extends TestCase
         $outcome = [
             ['the quick brown fox jumped over the lazy man sitting at a bus'
                 . ' stop drinking a can of coke'],
-            ['with a dandy umbrella'],
+            ['with a dandy   umbrella'],
             ['depuis qu’il avait emménagé à côté de chez elle, il y a de ça cinq ans.'],
             ['working with emoji 🤓'],
         ];